drivers:dac:ltc2672: Fix current_to_code api
Fix overflow of intermediate conversion value of
current in uA to code.

Fixes: 51f92f898c950b8e317cce2daca11da153ea6a10 ("dac: ltc2672: Remove float type usage")

+++ drivers/dac/ltc2672/ltc2672.c
@@ -155,7 +155,7 @@ uint32_t ltc2672_current_to_code(struct ltc2672_dev *device,
 		current_code = ((dac_current * LTC2672_12BIT_RESO) /
 				device->max_currents[out_ch]);
 	else
-		current_code = ((dac_current * LTC2672_16BIT_RESO) /
+		current_code = (((uint64_t)dac_current * LTC2672_16BIT_RESO) /
 				device->max_currents[out_ch]);
 
 	return current_code;
